Restructure cache decisions, move cache worker to own module
