+++ lib/apiCacheWorker/constants.ts
@@ -1,0 +1,3 @@
+// the threshold at which a cache becomes expired, relative to the "current"
+// time, in milliseconds.
+export const cacheExpirationThreshold = 1 * 60 * 1000; // 5 minutes
